Add EtcdResult.etcd_index; minor fixes.
We added EtcdResult.etcd_index and EtcdResult.raft_index as result of
parsing response headers from etcd, as requested in #39. Also, we fixed
the bug in the etcd.Client.write() response when updating a dir ( #38),
which is done by parsing the content of prevNode, which we formerly
discarded. Last, this patch makes etcd.Client.update() finally work
against directories, working around the unusual behaviour of the etcd
server in this respect.