Fix etcd.Client.machines
In our old version of the code, etcd.Client.api_execute called
Client.machines in case of failure, which called api_execute in
return. In the scenario of a client with a list of possible servers, and
the first of the list being down, we throw an exception right now,
because of this.

Using directly urllib3 and skipping api_execute in Client.machines
allows us to connect to the first available server.

Closes #51