Fix unhelpful log message when cluster ID changes
When the cluster ID changes, python-etcd writes a log message of
the form

    The UUID of the cluster changed from (expected) to (actual).

Since the expected cluster ID was overwritten with the actual value
immediately before this message was logged, the log message would
always use the same value in two places. This commit corrects the
message to use the correct value for (expected).