Moved to positional args in string formatting
In Python 3.1 the need for positional arguments in string formatting
became optional. However, there are a lot of people using Python 3.0.x
and 2.6.x who could benefit from using the python-etcd library. This
change would allow older versions of Python to be able to use the
library without requiring them to be officially supported.