Added respect for no_proxy.
This is rather important as the library will typically not allow
connection to localhost, when a proxy is otherwise set for the
system. It is common practise to avoid using a proxy for localhost
and 127.0.0.1.

This means that a local development version of etcd will work, if
you are otherwise using a proxy, if this patch isn't applied.