Reduce reraised-exception log to debug
When an exception occurs during a request, it can be for harmless
reasons (using an eventlet.timeout, for example).  Also, the exception
is already returned to the caller, who can better judge its severity.
Logging it here as an exception puts a lot of ERROR-level logs out
that might not, in fact, reflect errors,

This lowers the log level to a single-line debug.